fix types

+++ xtd.forms/include/xtd/forms/application.hpp
@@ -1,4 +1,5 @@
 #pragma once
+#include <cstdint>
 #include <memory>
 #include <vector>
 #include "application_informations.hpp"
@@ -77,7 +78,7 @@ namespace xtd {
       
     private:
       //static void wnd_proc(message& message);
-      static int wnd_proc_(intptr_t hwnd, int msg, intptr_t wparam, intptr_t lparam, intptr_t handle);
+      static intptr_t wnd_proc_(intptr_t hwnd, int32_t msg, intptr_t wparam, intptr_t lparam, intptr_t handle);
       static void wnd_proc(message& message);
       static void wm_activate_app(message& message);
       static void wm_enter_idle(message& message);
